requirements.txt
requests
python-telegram-bot[webhooks]==21.6
google-cloud-firestore